Tworzenie metody do przekazywania danych: Waluta żródłowa,docelowa i kwota. Tworzenie warunku kupna i sprzedaży waluty obcej

+++ src/main/java/com/example/currencyexchangerate/model/ExchangeRate.java
@@ -16,6 +16,6 @@ import java.math.BigDecimal;
 @NoArgsConstructor
 
 public class ExchangeRate {
-    private BigDecimal buy;
-    private BigDecimal sell;
+    private Double buy;
+    private Double sell;
 }
